Wrap templates not ready for grunt-init
Templates not ready for grunt-init are detected by looking for
a root/ and template.js file in the directory. If none are found,
the template is wrapped ready for grunt-init.

Also the template directory is installed in a lib/ directory
and a symbolic link is created in a bin/ directory pointing to
the directory in lib/.